Refactor: extract sort types to break circular deps
