+++ nsysu_selector_helper/src/services/index.ts
@@ -3,3 +3,4 @@ export * from './advancedFilterService';
 export * from './customQuickFiltersService';
 export * from './courseSortingService';
 export * from './filterPersistenceService';
+export * from '@/types/sorting';
